Move v5 ParameterSelect to "legacy" export
